% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ABX_2}
\alias{ABX_2}
\title{Another ABX looming data set supplied on 3 Dec}
\format{
\subsection{\code{ABX_2}}{

A data frame with 150 rows (each representing a time interval) and 12 columns, each column from the second column onward representing a subject
}
}
\usage{
ABX_2
}
\description{
An example dataset with the acceptable format
}
\keyword{datasets}
