% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ABX_2_avg_velocity}
\alias{ABX_2_avg_velocity}
\title{average_velocity vectors for ABX_2}
\format{
\subsection{\code{ABX_2_avg_velocity}}{

A data frame with 1 row and 12 columns, each column representing a subject
}
}
\usage{
ABX_2_avg_velocity
}
\description{
The average velocity vector associated with the ABX_Looming dataset
}
\keyword{datasets}
