% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ABX_Looming}
\alias{ABX_Looming}
\title{ABX looming data set}
\format{
\subsection{\code{ABX_Looming}}{

A data frame with 150 rows (each representing a time interval) and 6 columns, each column from the second column onward representing a subject
}
}
\usage{
ABX_Looming
}
\description{
An example dataset with the acceptable format
}
\keyword{datasets}
