% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ABX_Looming_baseline}
\alias{ABX_Looming_baseline}
\title{ABX looming data set average_velocity vectors}
\format{
\subsection{\code{ABX_Looming_baseline}}{

A data frame with 1 row and 5 columns, each column representing a subject
}
}
\usage{
ABX_Looming_baseline
}
\description{
The average velocity vector associated with the ABX_Looming dataset
}
\keyword{datasets}
