% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_time_plot.R
\name{accumulated_time_heat_plot}
\alias{accumulated_time_heat_plot}
\title{Function for creating heat plot for accumulated time}
\usage{
accumulated_time_heat_plot(
  cumu,
  master_cumulative_prof,
  time_to_consider = 1,
  experiment_time_in_sec = 15,
  subject_vec = NULL,
  show_values = FALSE
)
}
\description{
Function for creating heat plot for accumulated time
}
