% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_time_plot.R
\name{accumulated_time_plot}
\alias{accumulated_time_plot}
\title{Function for creating accumulated time plot}
\usage{
accumulated_time_plot(cumu, experiment_time_in_sec = 15, subject_vec = NULL)
}
\description{
Function for creating accumulated time plot
}
