% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_timepoints_new.R
\name{all_contig_sequence}
\alias{all_contig_sequence}
\title{helper functions for finding the index of all the elements that represent the start and the end of contiguous subsequences}
\usage{
all_contig_sequence(sequence)
}
\description{
helper functions for finding the index of all the elements that represent the start and the end of contiguous subsequences
}
