% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_time.R
\name{calculate_cumulative_time}
\alias{calculate_cumulative_time}
\title{Function for calculating cumulative time for a subject at a specific time for a particular event}
\usage{
calculate_cumulative_time(
  time,
  subject_considered,
  event_considered,
  cumulative_profile
)
}
\description{
Function for calculating cumulative time for a subject at a specific time for a particular event
}
