% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinedplot.R
\name{combinedplot}
\alias{combinedplot}
\title{A quick temporary function for generating combined trend and raster plot}
\usage{
combinedplot(data, avg_velocity, event.prof)
}
\arguments{
\item{data}{the data matrix; ABX_Looming for example}

\item{avg_velocity}{the average velocity vector; ABX_Looming-baseline for example}

\item{event.prof}{event profile generated from processed_eventprofiles(), the input of which is from EventProfiles(), the input of which is from P_DATA, which is output by lm_preprocess()}
}
\value{
a plot combining trend and raster plots
}
\description{
A quick temporary function for generating combined trend and raster plot
}
