% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumuprob_plot.R
\name{cumuprob_plot}
\alias{cumuprob_plot}
\title{A quick function for plotting a cumulative prob vs time plot; the definition of cumulative prob in this plot is to be given in a vignette}
\usage{
cumuprob_plot(event.prof, time_intervals)
}
\arguments{
\item{event.prof, }{the event profile generated via gutbehaviour::lm_preprocess(data, average.velocity) \%>\% EventProfiles(freeze_time_duration = 15, flight_time_duration = 3)}

\item{time_intervals, }{the time_interval vector of the P_DATA object returned from lm_preprocess()}
}
\value{
A list of two objects; it contains p, which is the plot.
}
\description{
A quick function for plotting a cumulative prob vs time plot; the definition of cumulative prob in this plot is to be given in a vignette
}
