% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_conditions.R
\name{.check_flight}
\alias{.check_flight}
\title{The flight condition supported should always satisfies the syntax
"> m * ave_velocity"}
\usage{
.check_flight(vec, ave_velocity)
}
\description{
The flight condition supported should always satisfies the syntax
"> m * ave_velocity"
}
