% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_conditions.R
\name{.check_flight2}
\alias{.check_flight2}
\title{The flight condition supported should always satisfies the syntax
"> m * ave_velocity"}
\usage{
.check_flight2(vec, ave_velocity, factor_flight)
}
\description{
The flight condition supported should always satisfies the syntax
"> m * ave_velocity"
}
