% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_conditions.R
\name{.check_freeze}
\alias{.check_freeze}
\title{Helper functions for checking whether a scalar fulfills the required condition for freezing and flight
The freeze condition supported should always satisfies the syntax
"< m * ave_velocity"}
\usage{
.check_freeze(vec, ave_velocity)
}
\arguments{
\item{vec}{the velocity row vector contained in the output tibble of lm_preprocess}

\item{ave_velocity}{an iterator object created from the average_velocity vector}
}
\description{
Helper functions for checking whether a scalar fulfills the required condition for freezing and flight
The freeze condition supported should always satisfies the syntax
"< m * ave_velocity"
}
