% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_timepoints.R
\name{.event_timepoints.helper}
\alias{.event_timepoints.helper}
\title{helper function for comparison}
\usage{
.event_timepoints.helper(array, comparend, Pointer, size)
}
\description{
helper function for comparison
}
