% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_timepoints.R
\name{.event_timepoints}
\alias{.event_timepoints}
\title{Given a numeric array containing only integers return the start and ending indices of all the possible continguous subsequences}
\usage{
.event_timepoints(array, m)
}
\arguments{
\item{array}{the array to be input}

\item{m}{a positive integer}
}
\value{
Record, a list of tuples (x,y) where x and y are the starting and the ending indices of any possible continguous subsequence
}
\description{
Given a numeric array containing only integers return the start and ending indices of all the possible continguous subsequences
}
