% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_timepoints.R
\name{.event_timepoints_all}
\alias{.event_timepoints_all}
\title{Given a numeric containing only integers return the start and ending indices of all the possible continguous subsequences (including that of only length 1)}
\usage{
.event_timepoints_all(array)
}
\arguments{
\item{array}{the array to be input}
}
\value{
Record, a list of tuples (x,y) where x and y are the starting and the ending indices of any possible continguous subsequence (including that of only length 1; in such case the indices are repeated)
}
\description{
Given a numeric containing only integers return the start and ending indices of all the possible continguous subsequences (including that of only length 1)
}
