% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_timepoints.R
\name{.event_timepoints_helper_all}
\alias{.event_timepoints_helper_all}
\title{helper function for .event_timepoints_all}
\usage{
.event_timepoints_helper_all(Pointer_a, Pointer_b, array, term_index)
}
\description{
helper function for .event_timepoints_all
}
