% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventProfiles.R
\name{.eventless_helper}
\alias{.eventless_helper}
\title{Helper function for identifying eventless profiles}
\usage{
.eventless_helper(result, i, condition = "freeze")
}
\description{
Helper function for identifying eventless profiles
}
