% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventProfiles.R
\name{.EventlessProfile}
\alias{.EventlessProfile}
\title{Helper function for identifying the time points at which there are no events}
\usage{
.EventlessProfile(ident, processed_data)
}
\description{
Helper function for identifying the time points at which there are no events
}
