% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventProfiles.R
\name{.FlightProfile}
\alias{.FlightProfile}
\title{Helper function for identifying flight condition}
\usage{
.FlightProfile(Match_flight, flight_time_duration)
}
\description{
Helper function for identifying flight condition
}
