% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventProfiles.R
\name{.FreezeProfile}
\alias{.FreezeProfile}
\title{Helper function for identifying freezing condition}
\usage{
.FreezeProfile(Match_freeze, freeze_time_duration)
}
\description{
Helper function for identifying freezing condition
}
