% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventProfiles.R
\name{.ShelterProfile}
\alias{.ShelterProfile}
\title{Helper function for identifying the timepoints of shelter condition}
\usage{
.ShelterProfile(Match_shelter, processed_data)
}
\description{
Helper function for identifying the timepoints of shelter condition
}
