% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Eventprofiles}
\alias{Eventprofiles}
\title{Event profile example}
\format{
\subsection{\code{Eventprofiles}}{

A data frame of dimension k*4, with each row representing an event of interest; the first 2 columns are the start and end times of the events respectively, while the third column is the name of the subject, the fourth column describes the event
}
}
\usage{
Eventprofiles
}
\description{
Using ABX_Looming and ABX_Looming_baseline as the input data, an event profile is generated using EventProfiles function
}
\keyword{datasets}
