% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventProfiles2.R
\name{EventProfiles2}
\alias{EventProfiles2}
\title{A function for creating event profiles given processed data from lm experiment, using specific check_conditions functions
allowing different freezing and flight conditions to be set}
\usage{
EventProfiles2(
  processed_data,
  freeze_time_duration,
  flight_time_duration,
  factor.freeze = 0.2,
  factor.flight = 3
)
}
\arguments{
\item{processed_data}{The list output by lm_preprocess}

\item{freeze_time_duration}{A positive integer indicating the minimum number of time intervals that should be spanned for a freeze event to be said to have occured}

\item{flight_time_duration}{A positive integer indicating the minimum number of time intervals that should be spanned for a flight event to be said to have occured}
}
\value{
EventProfiles, a data frame of dimension m*4; the first column is the start time of the event; second column the end time
}
\description{
A function for creating event profiles given processed data from lm experiment, using specific check_conditions functions
allowing different freezing and flight conditions to be set
}
