% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventProfiles2.R
\name{EventProfiles_all}
\alias{EventProfiles_all}
\title{A function for creating event profiles given processed data from lm experiment, using specific check_conditions functions. Moreover, sheltering events are also identified.}
\usage{
EventProfiles_all(
  processed_data,
  freeze_time_duration,
  flight_time_duration,
  factor.freeze = 0.2,
  factor.flight = 3
)
}
\description{
A function for creating event profiles given processed data from lm experiment, using specific check_conditions functions. Moreover, sheltering events are also identified.
}
