% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_timepoints_new.R
\name{filter_not_long}
\alias{filter_not_long}
\title{helper function for filter off elements that represent contiguous subsequences not long enough}
\usage{
filter_not_long(element, min_length = 0)
}
\description{
helper function for filter off elements that represent contiguous subsequences not long enough
}
