% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_time_plot.R
\name{generate_caption}
\alias{generate_caption}
\title{Helper function for generating caption in heat plot video}
\usage{
generate_caption(t)
}
\description{
Helper function for generating caption in heat plot video
}
