% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_time.R
\name{generate_cumulative_profiles}
\alias{generate_cumulative_profiles}
\title{Function for generating master cumulative time event profiles}
\usage{
generate_cumulative_profiles(P.DATA, event.prof)
}
\description{
Function for generating master cumulative time event profiles
}
