% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_time.R
\name{generate_cumutime_all_events}
\alias{generate_cumutime_all_events}
\title{Function for generating cumulative time profile at all considered times for all events}
\usage{
generate_cumutime_all_events(default_time_seq, P.DATA, cumulative_profile)
}
\description{
Function for generating cumulative time profile at all considered times for all events
}
