% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_time.R
\name{generate_cumutime_per_event}
\alias{generate_cumutime_per_event}
\title{Function for generating cumulative time profile at all considered times for a particular event}
\usage{
generate_cumutime_per_event(
  default_time_seq,
  event_considered,
  P.DATA,
  cumulative_profile
)
}
\description{
Function for generating cumulative time profile at all considered times for a particular event
}
