% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_time.R
\name{generate_cumutime_per_timestep}
\alias{generate_cumutime_per_timestep}
\title{Function for generating cumulative time profile at a particular time}
\usage{
generate_cumutime_per_timestep(
  time_considered,
  event_considered,
  P.DATA,
  cumulative_profile
)
}
\description{
Function for generating cumulative time profile at a particular time
}
