% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_time_plot.R
\name{generate_heat_plot_video}
\alias{generate_heat_plot_video}
\title{Function for generating heat plot video gif}
\usage{
generate_heat_plot_video(cumu_prof, video_name = "heat_plot_video.gif")
}
\description{
Function for generating heat plot video gif
}
