% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_time.R
\name{get_relevant_profile}
\alias{get_relevant_profile}
\title{Function for getting relevant profile from event profiles}
\usage{
get_relevant_profile(event.prof, subject_concerned, event_concerned)
}
\description{
Function for getting relevant profile from event profiles
}
