% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_preprocess.R
\name{lm_preprocess}
\alias{lm_preprocess}
\title{Preprocess the data from lm experiment}
\usage{
lm_preprocess(data, average_velocity)
}
\arguments{
\item{data}{A data frame that is the lm dataset of dimension m*(n+1); m is the number of time intervals and n is the number of subjects}

\item{average_velocity}{A data frame of dimension 1*n that contains the average velocities of the n subjects;}
}
\value{
A list containing the preprocessed data and other additional information
}
\description{
Preprocess the data from lm experiment
}
