% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_timepoints_new.R
\name{map_index_to_seq}
\alias{map_index_to_seq}
\title{helper functions for mapping index to sequence}
\usage{
map_index_to_seq(index, seq)
}
\description{
helper functions for mapping index to sequence
}
