% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulate_time_plot.R
\name{plot_subjects_trajectory}
\alias{plot_subjects_trajectory}
\title{Function for generating trajectory plot video gif}
\usage{
plot_subjects_trajectory(
  cumu_prof,
  subject_vec = NULL,
  video_name = "freeze_flight_trajectory.gif",
  experiment_time_in_sec = 15
)
}
\description{
Function for generating trajectory plot video gif
}
