% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{processed_data}
\alias{processed_data}
\title{Processed data}
\format{
\subsection{\code{processed_data}}{

A list containing the processed data. The first object in the list is the tibble (which should be the transpose of the original data)
}
}
\usage{
processed_data
}
\description{
A list of processed data after running lm_preprocess() on ABX_Looming and ABX_Looming_baseline
}
\keyword{datasets}
