% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{processed_event_data}
\alias{processed_event_data}
\title{processed event profile example}
\format{
\subsection{\code{processed_event_data}}{

A dataframe of dimension k*4 (the same as that of Eventprofiles); there should be no NA within processed_event_data
}
}
\usage{
processed_event_data
}
\description{
Using processed_eventprofiles function, an event profile that is Eventprofiles without the NAs is produced
}
\keyword{datasets}
