% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processed_eventprofiles.R
\name{processed_eventprofiles}
\alias{processed_eventprofiles}
\title{Processing event profiles generated from EventProfiles function for easier plotting}
\usage{
processed_eventprofiles(Eventprofiles)
}
\arguments{
\item{Eventprofiles}{the event profile generated by EventProfiles function. See documentation for EventProfiles for details.}
}
\value{
a data frame that contains only the complete cases (Eventprofiles without the NA) and formatting time
}
\description{
Processing event profiles generated from EventProfiles function for easier plotting
}
