% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendplot.R
\name{trendplot}
\alias{trendplot}
\title{A quick temporary function used for generating trend plot}
\usage{
trendplot(data, avg_velocity)
}
\arguments{
\item{data}{the data matrix, ABX_Looming for example}

\item{avg_velocity}{the average velocity vector, ABX_Looming_baseline for example}
}
\value{
a time-series line plot with smoothed curves
}
\description{
A quick temporary function used for generating trend plot
}
