% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various_lm_plots.R
\name{various_lm_plots}
\alias{various_lm_plots}
\title{Produce different types of plots for lm experiments}
\usage{
various_lm_plots(Eventprofiles, time = 0)
}
\arguments{
\item{Eventprofiles}{The event profile generated by EventProfiles function AND updated by processed_eventprofiles function}

\item{time}{a vector containing the times at which stimulus was given in the experiment; default is 0.}
}
\value{
a list of relevant plots
}
\description{
Produce different types of plots for lm experiments
}
