% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various_lm_plots.R
\name{various_lm_plots}
\alias{various_lm_plots}
\title{Produce different types of plots for lm experiments}
\usage{
various_lm_plots(event_profiles, time = 0, size_parameter = 10, time_intervals)
}
\arguments{
\item{event_profiles}{The event profile generated by EventProfiles function AND updated by processed_eventprofiles function}

\item{time}{a vector containing the times at which stimulus was given in the experiment; default is 0; the formats must be the same as the components of time_intervals vector (see the time_intervals argument below)}

\item{size_parameter}{a positive integer for setting the size of each event segment; the larger the integer the larger the segment}

\item{time_intervals}{the time interval vector in the processed data output by lm_preprocess}
}
\value{
a list of relevant plots, currenlt showing the raster plot first
}
\description{
Produce different types of plots for lm experiments
}
